bot/requirements.txt
pytelegrambotapi>=4.0.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
pinterest-dl>=1.0.0
Pillow>=10.0.0